+++ src/LDL/File/Finder/Adapter/Local/LocalFileFinderAdapter.php
@@ -7,6 +7,7 @@ use LDL\File\Finder\Adapter\AdapterInterface;
 use LDL\File\Finder\FoundFile;
 use LDL\Framework\Base\Collection\CallableCollectionInterface;
 use LDL\Validators\Chain\ValidatorChainInterface;
+use LDL\Validators\Collection\ValidatorCollectionInterface;
 use LDL\Validators\HasValidatorResultInterface;
 
 class LocalFileFinderAdapter implements AdapterInterface
@@ -17,9 +18,9 @@ class LocalFileFinderAdapter implements AdapterInterface
     private $validators;
 
     /**
-     * @var ValidatorChainInterface
+     * @var ValidatorCollectionInterface
      */
-    private $validatorsResult;
+    private $resultValidator;
 
     /**
      * @var CallableCollection
@@ -55,7 +56,12 @@ class LocalFileFinderAdapter implements AdapterInterface
         $this->onFile = $onFile ?? new CallableCollection();
         $this->onReject = $onReject ?? new CallableCollection();
         $this->onAccept = $onAccept ?? new CallableCollection();
-        $this->validatorsResult = $this->validators->filterByInterfaceRecursive(HasValidatorResultInterface::class);
+
+        $this->resultValidator = $this->validators->getChainItems()
+        ->getValidators()
+        ->filterByInterfaceRecursive(
+            HasValidatorResultInterface::class
+        );
     }
 
     public function find(iterable $directories, bool $recursive = true): iterable
@@ -100,7 +106,7 @@ class LocalFileFinderAdapter implements AdapterInterface
                     $foundFile = new FoundFile(
                         $file,
                         new \SplFileInfo($file),
-                        $this->validatorsResult->getCollection()
+                        $this->resultValidator
                     );
 
                     $this->onAccept->call($file, $this->validators);