+++ CHANGELOG.md
@@ -14,4 +14,5 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ### Changed
 
+- fixes/1200713875634268 - Fix to comply with validators (getChainItems)
 - fix/1200624131677614 - Fix validator - Remove config and dumpable. Add a description